@@ -9,7 +9,7 @@ NORMAL_JOIN += "LEFT JOIN spells AS recourse ON recourse.id = spells.recourse_id
 
 class SearchSpellsController < ApplicationController
 
-	before_filter :sanitize_params
+	before_action :sanitize_params
 
 protected
 	def valid_int?(i)
@@ -214,7 +214,7 @@ public
 		# For a given spell, pull in its classes
 		@spells.each do | spell |
 			@class = GetSpellClasses(spell.id)
-			spell["classes"] = @class.first["classes"]
+			spell.classes = @class.first["classes"]
 
 			# Also pull in effects and calculate results
 			@arrEffects,@hashStacks = GetSpellEffects(spell.id, spell.beneficial, spell.duration, spell.extra)
@@ -294,8 +294,9 @@ public
 			
 			# For a given spell, pull in its classes
 			@spells.each do | spell |
-				@class = MapSpellToCharClass.find_by_sql("SELECT group_concat(CONCAT(char_classes.name, '(', m.level, ')') SEPARATOR ', ') AS classes FROM char_classes INNER JOIN map_spell_to_char_classes as m ON m.class_id = char_classes.id WHERE m.spell_id = " + spell["id"].to_s + " ORDER BY level, name")
-				spell["classes"] = @class.first["classes"]
+				#@class = MapSpellToCharClass.find_by_sql("SELECT group_concat(CONCAT(char_classes.name, '(', m.level, ')') SEPARATOR ', ') AS classes FROM char_classes INNER JOIN map_spell_to_char_classes as m ON m.class_id = char_classes.id WHERE m.spell_id = " + spell["id"].to_s + " ORDER BY level, name")
+				@class = GetSpellClasses(spell.id)
+				spell.classes = @class.first["classes"]
 			end
 
 			if (@spells.length == 1)
